packages/observability/pyproject.toml
[project]
name = "sovereign-edge-observability"
version = "0.3.0"
requires-python = ">=3.11"
dependencies = [
    "structlog>=24.1,<26.0",
    "opentelemetry-sdk>=1.27,<2.0",
    "opentelemetry-exporter-otlp>=1.27,<2.0",
    "sovereign-edge-core",
]

[tool.uv.sources]
sovereign-edge-core = { workspace = true }

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/observability"]
